fix(frontend): isolate hire test timers and simplify onboarding components
