+++ autogpt_platform/frontend/src/app/(no-navbar)/onboarding/steps/ConnectStep/useConnectStep.ts
@@ -3,13 +3,12 @@
 import {
   getGetV2ListChatConnectionsQueryKey,
   useGetV2ListChatConnections,
-  useGetV2ListProviderModelTiers,
 } from "@/app/api/__generated__/endpoints/chat/chat";
 import { useQueryClient } from "@tanstack/react-query";
 import { useOAuthConnect } from "@/components/contextual/IntegrationsPanel/components/ConnectServiceDialog/components/DetailView/useOAuthConnect";
 
 import { useOnboardingWizardStore } from "../../store";
-import { hasLinkedSubscription, linkedModelsSentence } from "./helpers";
+import { hasLinkedSubscription } from "./helpers";
 
 export function useConnectStep() {
   const nextStep = useOnboardingWizardStore((s) => s.nextStep);
@@ -21,16 +20,6 @@ export function useConnectStep() {
   const offers =
     connectionsQuery.data?.status === 200
       ? connectionsQuery.data.data.offers
-      : undefined;
-
-  // What ChatGPT's tiers resolve to, which the connections list cannot answer
-  // for a connection the user has not made yet.
-  const tiersQuery = useGetV2ListProviderModelTiers({
-    query: { refetchOnWindowFocus: false },
-  });
-  const providers =
-    tiersQuery.data?.status === 200
-      ? tiersQuery.data.data.providers
       : undefined;
 
   // A successful sign-in does not move the wizard on: the box turns to
@@ -48,6 +37,5 @@ export function useConnectStep() {
     isConnecting: isPending,
     skip: nextStep,
     isAlreadyLinked: hasLinkedSubscription(offers),
-    models: linkedModelsSentence(providers),
   };
 }
